pyproject.toml
[project]
name = "pyshvtree"
version = "0.0.0"
description = "Python SHV Tree"
readme = "README.rst"
requires-python = ">=3.10"
dependencies = [
  "pyshv",
  "ruamel.yaml",
  "asyncinotify",
]

[project.urls]
"Homepage" = "https://gitlab.com/elektroline-predator/pyshvtree"
"Bug Tracker" = "https://gitlab.com/elektroline-predator/pyshvtree/-/issues"

[project.optional-dependencies]
docs = [
  "sphinx",
  "sphinx_rtd_theme",
  "sphinx-multiversion",
  "myst-parser",
]
test = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
]
lint = [
  "black",
  "isort",
  "pydocstyle",
  "pylint",
  "mypy",
]

[project.scripts]
shvtree-check = "shvtree.check.__main__:main"
shvtree-dummy = "shvtree.device.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["docs", "tests*"]
[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
omit = ["tests/*"]
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if typing.TYPE_CHECKING:"
]

[tool.isort]
profile = 'black'

[tool.pylint.main]
recursive = "y"
ignore = ["tests", "docs"]
fail-under = 9.5
[tool.pylint.format]
disable = [
  # Exception can be safely caught because there is BaseException for unsafe
  # ones (such as SystemExit) and there are reasons to catch all exceptions.
  "W0718",
  # Prevent usage of module name in documentation (seems unreasonable)
  "C0104",
  # We have pydocstyle for the following ones
  "C0103", "C0114", "C0115", "C0116",
]
# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
max-line-length = 88

[tool.mypy]
python_version = "3.10"
